Move slice_audio_file into tools module
